Support dry-running generators (#1028)
Running generators in dry-run mode, expect the GPDSC content to be
output to stdout. This allows consuming tools to parse the GPDSC and for
example resolve dependencies without any files being updated on disk.

The csolution tool doesn't currently use the dry-run output from the
generator. Since log messages from csolution and the generator also are
written to stdout, the generator need to enclose the GPDSC content
within special begin and end marks, so a tool running `csolution
--dry-run run` has to ignore everything outside those marks.

Begin and end marks are currently not fixed, but would preferably be
chosen to be consistent to avoid surprises.

Suggested begin mark: -----BEGIN GPDSC-----
Suggested end mark:   -----END GPDSC-----

Contributes to
[#206](https://github.com/Open-CMSIS-Pack/Open-CMSIS-Pack-Spec/issues/206).
Support for `additionalInputs` is left for later.

Spec update:
https://github.com/Open-CMSIS-Pack/Open-CMSIS-Pack-Spec/pull/242

Contributed by STMicroelectronics

+++ test/packs/ARM/RteTestGenerator/0.1.0/Generator/script.sh
@@ -3,7 +3,14 @@ echo $1
 echo $2 
 echo $3
 echo $4
+echo $5
 
-mkdir $1
-cp "$3/Templates/RteTest.gpdsc.template" "$1/RteTest.gpdsc"
-cp "$3/Templates/RteTest_Generated_Component.c.template" "$1/RteTest_Generated_Component.c"
+if [ "$5" = "--dry-run" ]; then
+  echo "-----BEGIN GPDSC-----"
+  cat "$3/Templates/RteTest.gpdsc.template"
+  echo "-----END GPDSC-----"
+else
+  mkdir $1
+  cp "$3/Templates/RteTest.gpdsc.template" "$1/RteTest.gpdsc"
+  cp "$3/Templates/RteTest_Generated_Component.c.template" "$1/RteTest_Generated_Component.c"
+fi
